Adiciona svelte e tailwindcss, atualiza dependenciais
