+++ index.html
@@ -1,23 +1,7 @@
 <!DOCTYPE html>
-    <html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Trabalho GPMS</title>
-        <style>
-            body {
-                display: flex;
-                flex-direction: column;
-                align-items: center;
-                background-color: gainsboro;
-            }
-        </style>
-    </head>
-    <body>
-        <div id="app">
-            <input type="checkbox" id="opcao_isometrico" name="opcao_isometrico"/>
-            <label>Ativar perspectiva isométrica</label>
-        </div>
-        <script type="module" src="/src/main.ts"></script>
-    </body>
+<html lang="en">
+	<head>
+		<script type="module" src="/src/main.ts"></script>
+	</head>
+	<body/>
 </html>
