+++ package.json
@@ -6,13 +6,19 @@
   "main": "src/main.ts",
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build",
+    "build": "vite build",
     "preview": "vite preview"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.194",
-    "typescript": "^5.0.2",
-    "vite": "^4.3.2"
+    "@sveltejs/vite-plugin-svelte": "^2.4.1",
+    "@types/lodash": "^4.14.195",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.24",
+    "svelte": "^3.59.1",
+    "svelte-preprocess": "^5.0.4",
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.1.3",
+    "vite": "^4.3.9"
   },
   "dependencies": {
     "lodash": "^4.17.21"
