+++ postcss.config.js
@@ -1,0 +1,10 @@
+export default {
+
+    plugins: {
+
+        tailwindcss: {},
+        autoprefixer: {},
+
+    }
+
+}
