+++ src/main.ts
@@ -1,31 +1,10 @@
-import { Jogo } from "./xadrez";
+import Main from "./Main.svelte";
+import "./style.css";
 
-// Define a classe Jogo como um elemento HTML <jogo-xadrez>
+const app: any = new Main({
 
-window.customElements.define("jogo-xadrez", Jogo);
+    target: document.body
 
-// Cria uma instância da classe Jogo e a insere na página
-
-let jogo = document.createElement("jogo-xadrez");
-document.body.appendChild(jogo);
-
-// Configuração do jogo
-
-let opcao_isometrico: HTMLInputElement = document.getElementsByName("opcao_isometrico")[0] as HTMLInputElement;
-opcao_isometrico.addEventListener("change", (event) => {
-    
-    if (event != null) {
-
-        if ((event.currentTarget as HTMLInputElement).checked) {
-
-            Jogo.isometrico = true;
-
-        } else {
-
-            Jogo.isometrico = false;
-
-        }
-
-    }
-    
 });
+
+export default app;
