+++ tailwind.config.js
@@ -1,0 +1,13 @@
+/** @type {import('tailwindcss').Config} */
+import * as Path from "path";
+
+export default {
+
+    content: [
+
+        Path.join("./src", "**/*.{html,js,svelte,ts}")
+
+    ]
+
+}
+
